requirements/development.txt
-r base.txt

# Dev tools
django-debug-toolbar==4.4.6
ipython==8.30.0
factory-boy==3.3.1
Faker==33.1.0
coverage==7.6.9
pytest-django==4.9.0
pytest-cov==6.0.0
